IAT-480: Refactor app root module to new directory structure.
