+++ src/app/item/components/item-history/item-history.component.spec.ts
@@ -2,7 +2,7 @@ import {async, ComponentFixture, TestBed} from "@angular/core/testing";
 import {APP_BASE_HREF} from "@angular/common";
 
 import {ItemHistoryComponent} from "./item-history.component";
-import {AppModule} from "../../../app.module";
+import {AppModule} from "../../../root/app.module";
 import {Logger} from "../../../core/logger.service/logger.service";
 import {AlertService} from "../../../core/alert.service/alert.service";
 import {ItemService} from "../../services/item.service";
