+++ src/app/item/components/load-item/load-item.component.spec.ts
@@ -1,6 +1,6 @@
 import {async, ComponentFixture, TestBed} from "@angular/core/testing";
 import {APP_BASE_HREF} from "@angular/common";
-import {AppModule} from "../../../app.module";
+import {AppModule} from "../../../root/app.module";
 import {Logger} from "../../../core/logger.service/logger.service";
 import {AlertService} from "../../../core/alert.service/alert.service";
 import { LoadItemComponent } from './load-item.component';
