+++ src/app/root/app.routing.ts
@@ -1,10 +1,10 @@
 import {NgModule} from "@angular/core";
 import {RouterModule, Routes} from "@angular/router";
-import {ItemDashboardComponent} from "./item/components/item-dashboard/item-dashboard.component";
-import {CreateItemSelectTypeComponent} from "./item/components/create-item-select-type/create-item-select-type.component";
-import {LoadItemComponent} from "./item/components/load-item/load-item.component";
-import {LoadItemRedirectComponent} from "./item/components/load-item-redirect/load-item-redirect.component";
-import {ResourceNotFoundComponent} from "./resource-not-found.component";
+import {ItemDashboardComponent} from "../item/components/item-dashboard/item-dashboard.component";
+import {CreateItemSelectTypeComponent} from "../item/components/create-item-select-type/create-item-select-type.component";
+import {LoadItemComponent} from "../item/components/load-item/load-item.component";
+import {LoadItemRedirectComponent} from "../item/components/load-item-redirect/load-item-redirect.component";
+import {ResourceNotFoundComponent} from "./components/resource-not-found.component/resource-not-found.component";
 
 const routes: Routes = [
   {path: '', component: ItemDashboardComponent},
