+++ src/app/root/components/app.component/app.component.html
@@ -1,7 +1,7 @@
-<!-- AlertService messages are displayed here -->
+<!-- AlertService messages -->
 <ng2-toasty [position]="'bottom-right'"></ng2-toasty>
 
-<!-- Loader messages are displayed here -->
+<!-- BusyService messages -->
 <busy-overlay></busy-overlay>
 
 <app-header></app-header>
