+++ src/app/root/components/app.component/app.component.spec.ts
@@ -8,18 +8,18 @@ import {ModalModule} from "ngx-bootstrap/modal";
 import {AlertModule} from "ngx-bootstrap/alert";
 
 import {AppComponent} from "./app.component";
-import {HeaderComponent} from "./header.component";
-import {FooterComponent} from "./footer.component";
-import {CreateItemSelectTypeComponent} from "./item/components/create-item-select-type/create-item-select-type.component";
-import {ItemDashboardComponent} from "./item/components/item-dashboard/item-dashboard.component";
-import {LoadSaItemComponent} from "./item/components/load-sa-item/load-sa-item.component";
+import {HeaderComponent} from "../header.component/header.component";
+import {FooterComponent} from "../footer.component/footer.component";
+import {CreateItemSelectTypeComponent} from "../../../item/components/create-item-select-type/create-item-select-type.component";
+import {ItemDashboardComponent} from "../../../item/components/item-dashboard/item-dashboard.component";
+import {LoadSaItemComponent} from "../../../item/components/load-sa-item/load-sa-item.component";
 import {ToastyModule} from "ng2-toasty";
-import {FindItemComponent} from "./item/components/find-item/find-item.component";
-import {Logger} from "./core/logger.service/logger.service";
-import {UserService} from "./core/user.service/user.service";
-import {AppInfoService} from "./core/app-info.service/app-info.service";
-import {CoreModule} from "./core/core.module";
-import {LoadWerItemComponent} from "./item/components/load-wer-item/load-wer-item.component";
+import {FindItemComponent} from "../../../item/components/find-item/find-item.component";
+import {Logger} from "../../../core/logger.service/logger.service";
+import {UserService} from "../../../core/user.service/user.service";
+import {AppInfoService} from "../../../core/app-info.service/app-info.service";
+import {CoreModule} from "../../../core/core.module";
+import {LoadWerItemComponent} from "../../../item/components/load-wer-item/load-wer-item.component";
 
 describe('AppComponent', () => {
   let component: AppComponent;
