+++ src/app/root/components/app.component/app.component.ts
@@ -2,8 +2,8 @@ import {Component} from "@angular/core";
 
 @Component({
   selector: 'app-root',
-  templateUrl: 'app.component.html',
-  styleUrls: ['.//app.component.less']
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.less']
 })
 export class AppComponent {
 }
