+++ src/app/root/components/header.component/header.component.spec.ts
@@ -4,9 +4,9 @@ import {HeaderComponent} from "./header.component";
 import {BsDropdownModule} from "ngx-bootstrap/dropdown";
 import {ModalModule} from "ngx-bootstrap/modal";
 import {CollapseDirective} from "ngx-bootstrap/collapse";
-import {Logger} from "./core/logger.service/logger.service";
-import {UserService} from "./core/user.service/user.service";
-import {AlertService} from "./core/alert.service/alert.service";
+import {Logger} from "../../../core/logger.service/logger.service";
+import {UserService} from "../../../core/user.service/user.service";
+import {AlertService} from "../../../core/alert.service/alert.service";
 import {ToastyConfig, ToastyService} from "ng2-toasty";
 
 describe('HeaderComponent', () => {
