+++ src/app/root/components/resource-not-found.component/resource-not-found.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ResourceNotFoundComponent } from './resource-not-found.component';
-import {AppModule} from "./app.module";
+import {AppModule} from "../../app.module";
 import {APP_BASE_HREF} from "@angular/common";
 
 describe('ResourceNotFoundComponent', () => {
